Replaced all axis=0 cases in 'localindex' with a call to Content::localindex_axis0.
